Out of tree build fixes
It looks like OpenPLi based kernel recipe builds out of the tree. This
confuses the Amlogic binary object magic for the NAND driver. The NAND
driver consists of some .z files which are copied to .o files during
build. If building out of tree this does not work. I disabled NAND
another solution is to patch the Makefiles to be out of source buildable.


Signed-off-by: graugans <k4230r6@gmail.com>